extract protobuf helpers from validation
