+++ ydb/public/lib/protobuf/ya.make
@@ -1,0 +1,11 @@
+LIBRARY()
+
+PEERDIR(
+    contrib/libs/protoc
+)
+
+SRCS(
+    helpers.cpp
+)
+
+END()
