+++ ydb/public/lib/validation/main.cpp
@@ -1,7 +1,6 @@
-#include "helpers.h"
-
-#include <ydb/public/api//protos/annotations/validation.pb.h>
 #include <ydb/library/yverify_stream/yverify_stream.h>
+#include <ydb/public/api/protos/annotations/validation.pb.h>
+#include <ydb/public/lib/protobuf/helpers.h>
 
 #include <google/protobuf/compiler/code_generator.h>
 #include <google/protobuf/compiler/plugin.h>
