+++ ydb/public/lib/validation/ya.make
@@ -3,10 +3,10 @@ PROGRAM()
 PEERDIR(
     contrib/libs/protoc
     ydb/public/api/protos/annotations
+    ydb/public/lib/protobuf
 )
 
 SRCS(
-    helpers.cpp
     main.cpp
 )
